Add edit functionality and form component

+++ frontend/src/app/pages/task/task.component.ts
@@ -8,17 +8,26 @@ import { TaskService } from '@lib/services/task/task.service';
 import { Task } from '@lib/interfaces/task.interface';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 
+import { FormComponent } from '@components/form/form.component';
+
 @Component({
   selector: 'app-task',
   standalone: true,
-  imports: [CommonModule, MatSlideToggleModule],
+  imports: [
+    CommonModule,
+    MatSlideToggleModule,
+    FormComponent
+  ],
   templateUrl: './task.component.html',
   styleUrl: './task.component.scss'
 })
 export class TaskComponent implements OnInit {
-  constructor(private store: Store<AppState>, private taskService: TaskService) { }
+  constructor(private store: Store<AppState>, public taskService: TaskService) { }
+  public selectedTask?: Task;
+  currentAction: 'list' | 'create' | 'edit' = 'list';
   allTasks$ = this.store.select(selectAllTasks);
   status$ = this.store.select((state) => state.tasks.status);
+
 
   ngOnInit(): void {
     // Fetch data from API
@@ -32,13 +41,33 @@ export class TaskComponent implements OnInit {
       });
   }
 
-  handleAddTask() {
-    let newTask: Task = { "name": "test", "description": "test", "status": "pedding" }
+  handleAddTask(newTask: Task) {
     this.taskService.addTask(newTask)
+    this.toggleActionTo('list');
+  }
+
+  handleEditTask(task: Task) {
+    this.taskService.updateTask(task)
+    this.toggleActionTo('list');
+  }
+
+  handleSetupEditForm(task: Task) {
+    this.selectedTask = task;
+    this.toggleActionTo('edit');
   }
 
   handleDeleteTask(task: Task) {
-    this.taskService.deleteTask(task)
+    if (confirm(`Are you sure you want to delete ${task.name} [${task._id}]?`))
+      this.taskService.deleteTask(task)
+
+  }
+
+  toggleActionTo(action: 'list' | 'create' | 'edit') {
+    this.currentAction = action;
+  }
+
+  toggleTo(action: any) {
+    this.toggleActionTo(action);
   }
 
 }
